+++ inst/examples/ex-resample.R
@@ -7,5 +7,11 @@ bootstrap(x, do = mean, n = 100)
 quant <- function(x) { quantile(x, probs = c(0.25, 0.75)) }
 bootstrap(x, n = 100, do = mean, f = quant)
 
+## Get the n bootstrap values
+bootstrap(x, n = 100, do = mean, f = function(x) { x })
+
 ## Jackknife
 jackknife(x, do = mean) # Sample mean
+
+## Get the leave-one-out values instead of summary
+jackknife(x, do = mean, f = function(x) { x })