+++ NEWS.md
@@ -3,6 +3,12 @@
 * Add `math_lcm()` and `math_gcd()` to compute the lowest common multiple and the greatest common divisor.
 * Add `interval_hdr()` and `interval_credible()` to compute the credible intervals.
 
+## Bugfixes & changes
+* `jackknife()` gained a new argument to apply a function on the leave-one-out values (`f`).
+
+## Internals
+* Add `with_seed()` to evaluate an expression with a temporarily seed.
+
 # arkhe 1.1.0
 ## New classes and methods
 * Add `needs()` to check for the availability of a package.